removed cast to int
git-svn-id: https://orca-robotics.svn.sf.net/svnroot/orca-robotics/orca/trunk@2968 534d5172-be3c-0410-9a3e-ccaa6372a113

+++ src/components/insgps/novatel/crc/ncrc32.cpp
@@ -138,7 +138,7 @@ int TestASCII( char* ucBuf_ )
 
    // Make a copy of the log with the '#' and '*' wrapper removed
    crcSize  = sizeof(ulCrc32);
-   dataSize = (int) szCheckCrc32 - (int) &ucBuf_[1]; 
+   dataSize = szCheckCrc32 - &ucBuf_[1]; 
    bufSize  = dataSize + crcSize;
    willy    = (unsigned char *) malloc(bufSize);
    memset(willy,0,bufSize);
